Keep emacs.d/init-loader/.gitignore

+++ emacs.d/create-symbolic-links.rb
@@ -12,6 +12,7 @@ def create_link(name, priority)
 end
 
 INIT_LOADER_DIR.each_entry do |entry|
+  next if ".gitignore" == entry
   path = (INIT_LOADER_DIR + entry)
   if path.file? or path.symlink?
     path.unlink
